requirements.txt
# Core Dependencies
google-generativeai>=0.3.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Authentication & Security
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# WhatsApp Integration
twilio>=8.10.0

# Database
aiosqlite>=0.19.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic[email]>=2.5.0

# Website Integration (Optional - install based on your needs)
# For MySQL database integration:
# mysql-connector-python>=8.0.0

# For PostgreSQL database integration:
# psycopg2-binary>=2.9.0

# For WooCommerce integration:
# woocommerce>=3.0.0

# HTTP requests for API integration
requests>=2.31.0
